Preprocessor ### What's done: Fixed after review

+++ save-preprocessor/src/main/kotlin/org/cqfn/save/preprocessor/controllers/DownloadProject.kt
@@ -4,6 +4,9 @@ import org.cqfn.save.preprocessor.Response
 import org.cqfn.save.repository.GitRepository
 
 import org.eclipse.jgit.api.Git
+import org.eclipse.jgit.api.errors.GitAPIException
+import org.eclipse.jgit.api.errors.InvalidRemoteException
+import org.eclipse.jgit.api.errors.TransportException
 import org.eclipse.jgit.transport.CredentialsProvider
 import org.eclipse.jgit.transport.UsernamePasswordCredentialsProvider
 import org.slf4j.LoggerFactory
@@ -35,9 +38,13 @@ class DownloadProject {
     fun upload(@RequestBody gitRepository: GitRepository): Response {
         val urlHash = gitRepository.url.hashCode()
         val tmpDir = File("$volumes/$urlHash")
-        tmpDir.deleteRecursively()
+        if (tmpDir.exists()) {
+            tmpDir.deleteRecursively()
+            log.info("For ${gitRepository.url} repository: dir $urlHash was deleted")
+        }
+        tmpDir.mkdirs()
         log.info("For ${gitRepository.url} repository: dir $urlHash was created")
-        val user = if (gitRepository.username != null && gitRepository.password != null) {
+        val userCredentials = if (gitRepository.username != null && gitRepository.password != null) {
             UsernamePasswordCredentialsProvider(gitRepository.username, gitRepository.password)
         } else {
             CredentialsProvider.getDefault()
@@ -45,16 +52,25 @@ class DownloadProject {
         try {
             Git.cloneRepository()
                 .setURI(gitRepository.url)
-                .setCredentialsProvider(user)
+                .setCredentialsProvider(userCredentials)
                 .setDirectory(tmpDir)
                 .call().use {
                     log.info("Repository cloned: ${gitRepository.url}")
                     // TODO post request to orchestrator
                     return Mono.just(ResponseEntity("Cloned", HttpStatus.ACCEPTED))
                 }
+        } catch (invalidRemoteException: InvalidRemoteException) {
+            log.warn("Invalidate remote exception while cloning ${gitRepository.url} repository", invalidRemoteException)
+            return Mono.just(ResponseEntity(invalidRemoteException.stackTraceToString(), HttpStatus.BAD_GATEWAY))
+        } catch (transportException: TransportException) {
+            log.warn("Transport operation failed while cloning ${gitRepository.url} repository", transportException)
+            return Mono.just(ResponseEntity(transportException.stackTraceToString(), HttpStatus.NOT_ACCEPTABLE))
+        } catch (gitAPIException: GitAPIException) {
+            log.warn("Error with git API while cloning ${gitRepository.url} repository", gitAPIException)
+            return Mono.just(ResponseEntity(gitAPIException.stackTraceToString(), HttpStatus.BAD_REQUEST))
         } catch (ex: Exception) {
             log.warn("Cloning ${gitRepository.url} repository failed", ex)
-            return Mono.just(ResponseEntity(ex.stackTraceToString(), HttpStatus.BAD_REQUEST))
+            return Mono.just(ResponseEntity(ex.stackTraceToString(), HttpStatus.INTERNAL_SERVER_ERROR))
         }
     }
 }
